+++ packages/cadl-ranch/src/actions/validate-mock-apis.ts
@@ -69,18 +69,8 @@ export async function validateMockApis({ scenariosPath }: ValidateMockApisConfig
 // THIS IS FOR BACK SUPPORT ONLY DO NOT ADD ANY NEW IGNORE. IMPLEMENT THE API.
 const NOT_IMPLEMENTED = [
   // Anne
-  "models/collections-basic",
-  "models/collections-models",
-  "models/enum-properties",
-  "models/inheritance",
-  "models/input-basic",
   "models/readonly-properties",
-  "models/roundtrip-basic",
-  "models/nested-models",
-  "models/primitive-properties",
-  "models/optional-properties",
-  "models/output-basic",
-  "clients/interfaces",
+  "clients/interfaces", // Need to figure out the client subclient operation group story for dpg.
 
   // Laurent
   "resiliency/dev-driven",