+++ packages/cadl-ranch/src/cli/cli.ts
@@ -3,10 +3,12 @@ import yargs from "yargs";
 import { validateScenarios } from "../actions/index.js";
 import { logger } from "../logger.js";
 import "source-map-support/register.js";
-import { serve } from "../actions/serve.js";
+import { serve, startInBackground, stop } from "../actions/serve.js";
 import { validateMockApis } from "../actions/validate-mock-apis.js";
 import { checkCoverage } from "../actions/check-coverage.js";
 import { generateScenarioSummary } from "../actions/generate-scenario-summary.js";
+import { uploadScenarioManifest } from "../actions/upload-scenario-manifest.js";
+import { uploadCoverageReport } from "../actions/upload-coverage-report.js";
 
 export const DEFAULT_PORT = 3000;
 
@@ -69,6 +71,51 @@ async function main() {
         });
       },
     )
+    .command("server", "Server management", (cmd) => {
+      cmd
+        .command(
+          "start <scenariosPath>",
+          "Start the server in the background.",
+          (cmd) => {
+            return cmd
+              .positional("scenariosPath", {
+                description: "Path to the scenarios and mock apis",
+                type: "string",
+                demandOption: true,
+              })
+              .option("port", {
+                alias: "p",
+                type: "number",
+                description: "Port where to host the server",
+                default: DEFAULT_PORT,
+              })
+              .option("coverageFile", {
+                type: "string",
+                description: "Path to the coverage file.",
+                default: join(process.cwd(), "cadl-ranch-coverage.json"),
+              });
+          },
+          async (args) =>
+            startInBackground({
+              scenariosPath: resolve(process.cwd(), args.scenariosPath),
+              port: args.port,
+              coverageFile: args.coverageFile,
+            }),
+        )
+        .command(
+          "stop",
+          "Stop the server running.",
+          (cmd) => {
+            return cmd.option("port", {
+              alias: "p",
+              type: "number",
+              description: "Port where to host the server",
+              default: DEFAULT_PORT,
+            });
+          },
+          async (args) => stop({ port: args.port }),
+        );
+    })
     .command(
       "serve <scenariosPath>",
       "Serve the mock api at the given paths.",
@@ -135,8 +182,8 @@ async function main() {
         await checkCoverage({
           scenariosPath: resolve(process.cwd(), args.scenariosPath),
           configFile: args.configFile,
-          mergedCoverageFile: args.mergedCoverageFile,
-          coverageFiles: args.coverageFiles,
+          mergedCoverageFile: resolve(process.cwd(), args.mergedCoverageFile),
+          coverageFiles: args.coverageFiles.map((x) => resolve(process.cwd(), x)),
           ignoreNotImplemented: args.ignoreNotImplemented,
         });
       },
@@ -157,6 +204,66 @@ async function main() {
         });
       },
     )
+    .command(
+      "upload-manifest <scenariosPath>",
+      "Upload the scenario manifest. DO NOT CALL in generator.",
+      (cmd) => {
+        return cmd
+          .positional("scenariosPath", {
+            description: "Path to the scenarios and mock apis",
+            type: "string",
+            demandOption: true,
+          })
+          .option("storageAccountName", {
+            type: "string",
+            description: "Name of the storage account",
+          })
+          .demandOption("storageAccountName");
+      },
+      async (args) => {
+        await uploadScenarioManifest({
+          scenariosPath: resolve(process.cwd(), args.scenariosPath),
+          storageAccountName: args.storageAccountName,
+        });
+      },
+    )
+    .command(
+      "upload-coverage",
+      "Upload the coverage report.",
+      (cmd) => {
+        return cmd
+          .option("coverageFile", {
+            type: "string",
+            description: "Path to the coverage file to upload.",
+            default: join(process.cwd(), "cadl-ranch-coverage.json"),
+          })
+          .demandOption("coverageFile")
+          .option("storageAccountName", {
+            type: "string",
+            description: "Name of the storage account",
+            default: join(process.cwd(), "cadl-ranch-coverage.json"),
+          })
+          .demandOption("storageAccountName")
+          .option("generatorName", {
+            type: "string",
+            description: "Name of generator",
+          })
+          .demandOption("generatorName")
+          .option("generatorVersion", {
+            type: "string",
+            description: "Version of generator",
+          })
+          .demandOption("generatorVersion");
+      },
+      async (args) => {
+        await uploadCoverageReport({
+          coverageFile: resolve(process.cwd(), args.coverageFile),
+          storageAccountName: args.storageAccountName,
+          generatorName: args.generatorName,
+          generatorVersion: args.generatorVersion,
+        });
+      },
+    )
     .demandCommand(1, "You must use one of the supported commands.")
     .parse();
 }
@@ -166,3 +273,8 @@ main().catch((error) => {
   console.log("Error", error);
   process.exit(1);
 });
+
+process.on("SIGTERM", () => process.exit(2));
+process.on("SIGINT", () => process.exit(2));
+process.on("SIGUSR1", () => process.exit(2));
+process.on("SIGUSR2", () => process.exit(2));