+++ packages/cadl-ranch-specs/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @azure-tools/cadl-ranch-specs
 
+## 0.3.2
+
+### Patch Changes
+
+- [#125](https://github.com/Azure/cadl-ranch/pull/125) [`78f52a6`](https://github.com/Azure/cadl-ranch/commit/78f52a69c09af377a1e0d4d9592cd17207982283) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Redresign Models usage scenarios
+
+- Updated dependencies [[`fc7ac81`](https://github.com/Azure/cadl-ranch/commit/fc7ac81e7bb6c8be4f5aa01863ab7f811dc96c52), [`78f52a6`](https://github.com/Azure/cadl-ranch/commit/78f52a69c09af377a1e0d4d9592cd17207982283)]:
+  - @azure-tools/cadl-ranch@0.2.3
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`4634cd7`](https://github.com/Azure/cadl-ranch/commit/4634cd7aa834b44bf58695a8a5c0b090ce0c525d), [`23665a0`](https://github.com/Azure/cadl-ranch/commit/23665a0a9abd1428486c274d5b72971ee1f8c379)]:
+  - @azure-tools/cadl-ranch@0.2.2
+
+## 0.3.0
+
+### Minor Changes
+
+- [#105](https://github.com/Azure/cadl-ranch/pull/105) [`08f06ba`](https://github.com/Azure/cadl-ranch/commit/08f06ba886189114fde67c68ccc433d3381cf5d9) Thanks [@msyyc](https://github.com/msyyc)! - Adding specs and mockapi to support basic lro
+
+### Patch Changes
+
+- [#118](https://github.com/Azure/cadl-ranch/pull/118) [`a307dea`](https://github.com/Azure/cadl-ranch/commit/a307deab3c1af32935992bd2fd8340f5879aeec5) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Fix: should be allowed on interface and namespace as well(next to @scenario)
+
+- Updated dependencies [[`a307dea`](https://github.com/Azure/cadl-ranch/commit/a307deab3c1af32935992bd2fd8340f5879aeec5)]:
+  - @azure-tools/cadl-ranch-expect@0.1.10
+  - @azure-tools/cadl-ranch@0.2.1
+
+## 0.2.0
+
+### Minor Changes
+
+- [#111](https://github.com/Azure/cadl-ranch/pull/111) [`b58be0e`](https://github.com/Azure/cadl-ranch/commit/b58be0e35aae20dad18f6c50f1b9ec407ca86e18) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Upload scenario manifest to storage account for dashboard
+
+### Patch Changes
+
+- [#113](https://github.com/Azure/cadl-ranch/pull/113) [`1fc9000`](https://github.com/Azure/cadl-ranch/commit/1fc90008a87acead33c86567c167dbcaa524115f) Thanks [@haolingdong-msft](https://github.com/haolingdong-msft)! - bug fix, update models/property-optional /duration operation
+
+- [#107](https://github.com/Azure/cadl-ranch/pull/107) [`9d6ae37`](https://github.com/Azure/cadl-ranch/commit/9d6ae3770ef6da76172377a922b5b6a52ea55d28) Thanks [@changlong-liu](https://github.com/changlong-liu)! - Fix route in dictionary mock apis.
+
+- [#114](https://github.com/Azure/cadl-ranch/pull/114) [`f34dc29`](https://github.com/Azure/cadl-ranch/commit/f34dc29d87003d6a3e5c37b78e2e9d83cabc27d7) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Upload coverage report and show aggregated report on dashboard
+
+- [#109](https://github.com/Azure/cadl-ranch/pull/109) [`069a616`](https://github.com/Azure/cadl-ranch/commit/069a616b32b3f898b5396f0a592ea53620cf571b) Thanks [@iscai-msft](https://github.com/iscai-msft)! - fix coverage report for invalid authentication tests
+
+- Updated dependencies [[`b58be0e`](https://github.com/Azure/cadl-ranch/commit/b58be0e35aae20dad18f6c50f1b9ec407ca86e18), [`f34dc29`](https://github.com/Azure/cadl-ranch/commit/f34dc29d87003d6a3e5c37b78e2e9d83cabc27d7)]:
+  - @azure-tools/cadl-ranch@0.2.0
+
+## 0.1.12
+
+### Patch Changes
+
+- [#108](https://github.com/Azure/cadl-ranch/pull/108) [`69518f3`](https://github.com/Azure/cadl-ranch/commit/69518f30d90cfdbae4ec8c7493229a2c2fe4fa01) Thanks [@iscai-msft](https://github.com/iscai-msft)! - Fix model type for collections model optional property testing
+
+- [#91](https://github.com/Azure/cadl-ranch/pull/91) [`cc4dd0f`](https://github.com/Azure/cadl-ranch/commit/cc4dd0fbeb58e7e85fd5ec6282739c49338daa71) Thanks [@pshao25](https://github.com/pshao25)! - Fix datetime compare issue
+
+- [#104](https://github.com/Azure/cadl-ranch/pull/104) [`e622348`](https://github.com/Azure/cadl-ranch/commit/e6223481a78f7b9844fee6d9457d01d9adf05860) Thanks [@changlong-liu](https://github.com/changlong-liu)! - Adding specs and mockapi for dictionary.
+
+- [#74](https://github.com/Azure/cadl-ranch/pull/74) [`cf1b966`](https://github.com/Azure/cadl-ranch/commit/cf1b9660e195ffca5f4292830f3e78f7af748534) Thanks [@changlong-liu](https://github.com/changlong-liu)! - Add specs for inheritance models with or without discriminator.
+
 ## 0.1.11
 
 ### Patch Changes