+++ packages/cadl-ranch/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @azure-tools/cadl-ranch
 
+## 0.2.3
+
+### Patch Changes
+
+- [#123](https://github.com/Azure/cadl-ranch/pull/123) [`fc7ac81`](https://github.com/Azure/cadl-ranch/commit/fc7ac81e7bb6c8be4f5aa01863ab7f811dc96c52) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Cleanup no implemented mock api list
+
+- [#125](https://github.com/Azure/cadl-ranch/pull/125) [`78f52a6`](https://github.com/Azure/cadl-ranch/commit/78f52a69c09af377a1e0d4d9592cd17207982283) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Redresign Models usage scenarios
+
+## 0.2.2
+
+### Patch Changes
+
+- [#119](https://github.com/Azure/cadl-ranch/pull/119) [`4634cd7`](https://github.com/Azure/cadl-ranch/commit/4634cd7aa834b44bf58695a8a5c0b090ce0c525d) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Not supported should be a supported coverage status
+
+- [#122](https://github.com/Azure/cadl-ranch/pull/122) [`23665a0`](https://github.com/Azure/cadl-ranch/commit/23665a0a9abd1428486c274d5b72971ee1f8c379) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Save coverage on process exiting only.
+
+## 0.2.1
+
+### Patch Changes
+
+- [#118](https://github.com/Azure/cadl-ranch/pull/118) [`a307dea`](https://github.com/Azure/cadl-ranch/commit/a307deab3c1af32935992bd2fd8340f5879aeec5) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Fix: should be allowed on interface and namespace as well(next to @scenario)
+
+- Updated dependencies [[`a307dea`](https://github.com/Azure/cadl-ranch/commit/a307deab3c1af32935992bd2fd8340f5879aeec5)]:
+  - @azure-tools/cadl-ranch-expect@0.1.10
+
+## 0.2.0
+
+### Minor Changes
+
+- [#111](https://github.com/Azure/cadl-ranch/pull/111) [`b58be0e`](https://github.com/Azure/cadl-ranch/commit/b58be0e35aae20dad18f6c50f1b9ec407ca86e18) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Upload scenario manifest to storage account for dashboard
+
+### Patch Changes
+
+- [#114](https://github.com/Azure/cadl-ranch/pull/114) [`f34dc29`](https://github.com/Azure/cadl-ranch/commit/f34dc29d87003d6a3e5c37b78e2e9d83cabc27d7) Thanks [@timotheeguerin](https://github.com/timotheeguerin)! - Upload coverage report and show aggregated report on dashboard
+
+- Updated dependencies [[`f34dc29`](https://github.com/Azure/cadl-ranch/commit/f34dc29d87003d6a3e5c37b78e2e9d83cabc27d7)]:
+  - @azure-tools/cadl-ranch-coverage-sdk@0.1.1
+
 ## 0.1.9
 
 ### Patch Changes